pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = ""
authors = ["aisaev <an.isaev7@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.7"
django-environ = "^0.10.0"
gunicorn = "^20.1.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
social-auth-app-django = "^5.2.0"
django-filter = "^23.1"
pydantic = "^1.10.7"
requests = "^2.30.0"


[tool.poetry.group.dev.dependencies]
django-extensions = "^3.2.1"
pre-commit = "^3.2.2"
mypy = "^1.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-lenght=120
target-version=["py310"]
skip-string-normalization=true
skip-magic-trailing-coma=true
exclude ='''
(
  /(
      \.git
    | \.github
    | \.venv
  )/
)
'''
